Move vendor deps to its own plugin
